Use int64_t instead of long long

+++ libmonit/src/system/Time.c
@@ -1118,15 +1118,15 @@ char *Time_uptime(time_t sec, char *result) {
                 result[0] = 0;
                 if (sec > 0) {
                         if ((r = sec/86400) > 0) {
-                                n = snprintf(result, 24, "%lldd", (long long)r);
+                                n = snprintf(result, 24, "%lldd", (int64_t)r);
                                 sec -= r * 86400;
                         }
                         if ((r = sec/3600) > 0) {
-                                n += snprintf(result + n, (24 - n), "%s%lldh", n ? ", " : "", (long long)r);
+                                n += snprintf(result + n, (24 - n), "%s%lldh", n ? ", " : "", (int64_t)r);
                                 sec -= r * 3600;
                         }
                         r = sec/60;
-                        snprintf(result + n, (24 - n), "%s%lldm", n ? ", " : "", (long long)r);
+                        snprintf(result + n, (24 - n), "%s%lldm", n ? ", " : "", (int64_t)r);
                 }
         }
         return result;
